requirements-gpu-win311.txt
fastapi==0.115.5
uvicorn==0.30.6
requests==2.32.3
python-dotenv==1.0.1
openai>=1.40.0,<2

opencv-python==4.12.0.88
numpy==2.1.3
pillow==11.0.0
reportlab==4.2.5
python-osc==1.8.3

soundfile==0.13.1
sounddevice==0.4.7

ultralytics==8.3.25

panda3d==1.10.13
panda3d-gltf==1.2.1
# (omitimos simplepbr en 3.11)

Cython==3.0.10
TTS==0.21.3
